[misc] Rename module
